requirements.txt
Django==4.2.13
djangorestframework==3.15.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1

